Switched to official OS, raspberryos.
The changes are more or less cosmetic, as far as actual meaningful content, but were necessary to get the tuner to work.
The alsa device name changed slightly, all shared libraries had to be put after the objects to avoid unresolved symbol errors, the version of armpl changed, etc. etc.